requirements.txt
faiss-cpu>=1.8.0
numpy>=1.26.0
scikit-learn>=1.4.0
sentence-transformers>=2.7.0
openai>=1.30.0
